+++ src/loader/loader.mli
@@ -19,38 +19,38 @@
 (*                                                                        *)
 (**************************************************************************)
 
-type ('a, 'b, 'c) t_pack = ELF of 'a | PE of 'b | Dump of 'c
+type ('a, 'b, 'c) t_pack = ELF of 'a | PE of 'b | Raw of 'c
 
 type ('a, 'b, 'c) header_pack =
   | ELF_header of 'a
   | PE_header of 'b
-  | Dump_header of 'c
+  | Raw_header of 'c
 
 include
   Loader_sigs.S
     with type Section.t =
           ( Loader_elf.Section.t,
             Loader_pe.Section.t,
-            Loader_dump.Section.t )
+            Loader_raw.Section.t )
           t_pack
      and type Symbol.t =
-          (Loader_elf.Symbol.t, Loader_pe.Symbol.t, Loader_dump.Symbol.t) t_pack
+          (Loader_elf.Symbol.t, Loader_pe.Symbol.t, Loader_raw.Symbol.t) t_pack
      and type Img.t =
-          (Loader_elf.Img.t, Loader_pe.Img.t, Loader_dump.Img.t) t_pack
+          (Loader_elf.Img.t, Loader_pe.Img.t, Loader_raw.Img.t) t_pack
      and type Section.header =
           ( Loader_elf.Section.header,
             Loader_pe.Section.header,
-            Loader_dump.Section.header )
+            Loader_raw.Section.header )
           header_pack
      and type Symbol.header =
           ( Loader_elf.Symbol.header,
             Loader_pe.Symbol.header,
-            Loader_dump.Symbol.header )
+            Loader_raw.Symbol.header )
           header_pack
      and type Img.header =
           ( Loader_elf.Img.header,
             Loader_pe.Img.header,
-            Loader_dump.Img.header )
+            Loader_raw.Img.header )
           header_pack
 
 module View : sig